feat: re-home some utility functions
